# openapi.yml
openapi: '3.0.2'
info:
  title: gitploy.
  version: '1.0'
servers:
  - url: http://localhost/api/v1
  - url: https://gitploy.jp.ngrok.io/api/v1
paths:
...
  /repos:
    get:
      tags:
        - Repo
      summary: List repositories the user can access.
      parameters:
        - in: query
          name: sort 
          schema:
            type: boolean
            default: false
        - in: query
          name: q
          schema:
            type: string
            default: ""
        - in: query
          name: namespace
          schema:
            type: string
            default: ""
        - in: query
          name: name
          schema:
            type: string
            default: ""
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            description: The page number
        - in: query
          name: per_page
          schema:
            type: integer
            default: 30
            description: The number per page
      responses:
        '200':
          description: Repositories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repositories'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401': 
          $ref: '#/components/responses/401Unauthorized'
        '402':
          $ref: '#/components/responses/402PaymentRequired'
        '500':
          $ref: '#/components/responses/500InternalError'
  /repos/{id}:
    get:
      tags:
        - Repo
      summary: Get the repository.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            description: The repository id
      responses:
        '200':
          description: Repository
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '401': 
          $ref: '#/components/responses/401Unauthorized'
        '402':
          $ref: '#/components/responses/402PaymentRequired'
        '403': 
          $ref: '#/components/responses/403Forbidden'
        '500':
          $ref: '#/components/responses/500InternalError'
    patch:
      tags:
        - Repo
      summary: Update the repository.
      description: Basically patch the respository field. And it creates a new webhook when it activates.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                config_path:
                  type: string
                  default: deploy.yml
                active:
                  type: boolean
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            description: The repository id
      responses:
        '200':
          description: Ok
        '401': 
          $ref: '#/components/responses/401Unauthorized'
        '402':
          $ref: '#/components/responses/402PaymentRequired'
        '403': 
          $ref: '#/components/responses/403Forbidden'
        '500':
          $ref: '#/components/responses/500InternalError'
...
security:
  - BearerAuth: []
  - OAuth2:
    - read
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
        avatar:
          type: string
        admin:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
        edges:
          type: object
          properties:
            chat_user:
              $ref: '#/components/schemas/ChatUser'
      required:
        - id
        - login
        - admin
        - avatar
        - created_at
        - updated_at
        - edges
    ChatUser:
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - created_at
        - updated_at
    Repositories:
      type: array
      items:
        $ref: '#/components/schemas/Repository'
    Repository:
      type: object
      properties:
        id:
          type: string
        namespace:
          type: string
          description: Repository owner
        name:
          type: string
        description:
          type: string
        config_path:
          type: string
        active:
          type: boolean
        webhook_id:
          type: integer
        locked:
          type: boolean
        synced_at:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        latest_deployed_at:
          type: string
        edges:
          type: object
          properties:
            deployments:
              $ref: '#/components/schemas/Deployments'
      required:
        - id
        - namespace
        - name
        - description
        - config_path
        - active
        - locked
        - created_at
        - updated_at
...
    Deployments:
      type: array
      items:
        $ref: '#/components/schemas/Deployment'
    Deployment:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum:
            - commit
            - branch
            - tag
        ref:
          type: string
        env:
          type: string
        status:
          type: string
          enum:
            - waiting
            - created
            - running
            - success
            - failure
            - canceled
        uid:
          type: integer
        sha:
          type: string
        html_url:
          type: string
        is_rollback:
          type: boolean
        is_approval_enabled:
          type: boolean
        required_approval_count:
          type: number
        created_at:
          type: string
        updated_at:
          type: string
        edges:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            repo:
              $ref: '#/components/schemas/Repository'
            deployment_statuses:
              type: array
              items:
                $ref: '#/components/schemas/DeploymentStatus'
      required:
        - id
        - type
        - ref
        - env
        - status
        - is_rollback
        - is_approval_enabled
        - required_approval_count
        - created_at
        - updated_at
    DeploymentStatus:
      type: object
      properties:
        id:
          type: number
        status:
          type: string
        description:
          type: string
        log_url:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - status
        - created_at
        - updated_at
...
    Error:
      type: object
      properties:
        message: 
          type: string
      required:
        - message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Hash key of user
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://github.com/oauth/authorize
          tokenUrl: https://github.com/oauth/token
          scopes:
            read: Grants read access
            write: Grants write access
            admin: Grants access to admin operations
  responses:
    400BadRequest:
      description: |
        The request could not be understood by the server due to malformed syntax.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    401Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    402PaymentRequired:
      description: License is expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    403Forbidden:
      description: Permisson denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    500InternalError:
      description: Server internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'